Load configuratin from config.py
