+++ requirements.txt
@@ -1,6 +1,8 @@
 Flask==2.1.1
 Flask-Cors==3.0.10
+flask-sqlalchemy
+gunicorn==20.1.0
+loguru==0.6.0
+pandas==1.4.2
 pandas==1.4.2
 psycopg2==2.9.3
-SQLAlchemy==1.4.35
-pandas==1.4.2
